Push a bunch of akka-http code into the shared lib
