+++ storage/ingests_api/src/main/scala/uk/ac/wellcome/platform/storage/ingests/api/IngestsApi.scala
@@ -4,14 +4,13 @@ import java.net.URL
 
 import akka.actor.ActorSystem
 import akka.http.scaladsl.Http
-import akka.http.scaladsl.server.{ExceptionHandler, RejectionHandler}
 import akka.stream.ActorMaterializer
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB
 import grizzled.slf4j.Logging
 import uk.ac.wellcome.Runnable
 import uk.ac.wellcome.messaging.sns.SNSWriter
 import uk.ac.wellcome.platform.archive.common.config.models.HTTPServerConfig
-import uk.ac.wellcome.platform.archive.common.http.HttpMetrics
+import uk.ac.wellcome.platform.archive.common.http.{HttpMetrics, WellcomeHttpApp}
 import uk.ac.wellcome.platform.archive.common.progress.monitor.ProgressTracker
 import uk.ac.wellcome.storage.dynamo.DynamoConfig
 
@@ -45,20 +44,9 @@ class IngestsApi(
     contextURL = contextURL
   )
 
-  implicit val rejectionHandler: RejectionHandler = router.rejectionHandler
-  implicit val exceptionHandler: ExceptionHandler = router.exceptionHandler
-  val bindingFuture: Future[Http.ServerBinding] = Http()
-    .bindAndHandle(
-      handler = router.routes,
-      interface = httpServerConfig.host,
-      port = httpServerConfig.port
-    )
-
   def run(): Future[Http.HttpTerminated] =
-    bindingFuture
-      .map(b => {
-        info(s"Listening on ${httpServerConfig.host}:${httpServerConfig.port}")
-        b
-      })
-      .flatMap { _.whenTerminated }
+    WellcomeHttpApp.run(
+      httpServerConfig = httpServerConfig,
+      router = router
+    )
 }
