+++ storage/ingests_api/src/main/scala/uk/ac/wellcome/platform/storage/ingests/api/Router.scala
@@ -3,23 +3,14 @@ package uk.ac.wellcome.platform.storage.ingests.api
 import java.net.URL
 import java.util.UUID
 
-import akka.http.scaladsl.marshalling.Marshal
-import akka.http.scaladsl.model._
 import akka.http.scaladsl.model.StatusCodes._
 import akka.http.scaladsl.model.headers.Location
 import akka.http.scaladsl.server._
-import akka.stream.scaladsl.Flow
-import akka.util.ByteString
-import grizzled.slf4j.Logging
-import io.circe.{CursorOp, Printer}
 import uk.ac.wellcome.platform.archive.common.config.models.HTTPServerConfig
-import uk.ac.wellcome.platform.archive.common.http.HttpMetrics
 import uk.ac.wellcome.platform.archive.common.http.models.ErrorResponse
+import uk.ac.wellcome.platform.archive.common.http.{BaseRouter, HttpMetrics}
 import uk.ac.wellcome.platform.archive.common.models.StorageSpace
-import uk.ac.wellcome.platform.archive.common.models.bagit.{
-  BagId,
-  ExternalIdentifier
-}
+import uk.ac.wellcome.platform.archive.common.models.bagit.{BagId, ExternalIdentifier}
 import uk.ac.wellcome.platform.archive.common.progress.models.Progress
 import uk.ac.wellcome.platform.archive.common.progress.monitor.ProgressTracker
 import uk.ac.wellcome.platform.archive.display.{
@@ -32,23 +23,14 @@ import scala.concurrent.ExecutionContext
 
 class Router(progressTracker: ProgressTracker,
              progressStarter: ProgressStarter,
-             httpMetrics: HttpMetrics,
+             val httpMetrics: HttpMetrics,
              httpServerConfig: HTTPServerConfig,
-             contextURL: URL)(implicit ec: ExecutionContext)
-    extends Logging {
+             val contextURL: URL)(implicit val ec: ExecutionContext)
+    extends BaseRouter {
 
   import akka.http.scaladsl.server.Directives._
   import de.heikoseeberger.akkahttpcirce.ErrorAccumulatingCirceSupport._
   import uk.ac.wellcome.json.JsonUtil._
-
-  implicit val printer = Printer.noSpaces.copy(dropNullValues = true)
-
-  import uk.ac.wellcome.platform.archive.display.DisplayProvider._
-
-  val sendCloudWatchMetrics: Directive0 = mapResponse { resp: HttpResponse =>
-    httpMetrics.sendMetric(resp)
-    resp
-  }
 
   def routes: Route = sendCloudWatchMetrics {
     pathPrefix("progress") {
@@ -109,92 +91,6 @@ class Router(progressTracker: ProgressTracker,
     }
   }
 
-  def rejectionHandler: RejectionHandler =
-    RejectionHandler
-      .newBuilder()
-      .handle {
-        case MalformedRequestContentRejection(_, causes: DecodingFailures) =>
-          handleDecodingFailures(causes)
-      }
-      .result()
-      .seal
-      .mapRejectionResponse {
-        case res @ HttpResponse(
-              statusCode,
-              _,
-              HttpEntity.Strict(contentType, _),
-              _) if contentType != ContentTypes.`application/json` =>
-          transformToJsonErrorResponse(statusCode, res)
-        case x => x
-      }
-      .mapRejectionResponse { resp: HttpResponse =>
-        httpMetrics.sendMetric(resp)
-        resp
-      }
-
-  implicit def exceptionHandler: ExceptionHandler = ExceptionHandler {
-    case err: Exception =>
-      logger.error(s"Unexpected exception $err")
-      val error = ErrorResponse(
-        context = contextURL,
-        statusCode = InternalServerError,
-        description = err.toString
-      )
-      httpMetrics.sendMetricForStatus(InternalServerError)
-      complete(InternalServerError -> error)
-  }
-
   private def createLocationHeader(progress: Progress) =
     Location(s"${httpServerConfig.externalBaseURL}/${progress.id}")
-
-  private def handleDecodingFailures(causes: DecodingFailures) = {
-    val message = causes.failures.map { cause =>
-      val path = CursorOp.opsToPath(cause.history)
-
-      // Error messages returned by Circe are somewhat inconsistent and we also return our
-      // own error messages when decoding enums (DisplayIngestType and DisplayStorageProvider).
-      val reason = cause.message match {
-        // "Attempt to decode value on failed cursor" seems to mean in circeworld
-        // that a required field was not present.
-        case s if s.contains("Attempt to decode value on failed cursor") =>
-          "required property not supplied."
-        // These are errors returned by our custom decoders for enum.
-        case s if s.contains("valid values") => s
-        // If a field exists in the JSON but it's of the wrong format
-        // (for example the schema says it should be a String but an object has
-        // been supplied instead), the error message returned by Circe only
-        // contains the expected type.
-        case s => s"should be a $s."
-      }
-
-      s"Invalid value at $path: $reason"
-    }
-
-    complete(
-      BadRequest -> ErrorResponse(
-        context = contextURL,
-        statusCode = BadRequest,
-        description = message.toList.mkString("\n")
-      )
-    )
-  }
-
-  private def transformToJsonErrorResponse(statusCode: StatusCode,
-                                           res: HttpResponse) = {
-    val errorResponseMarshallingFlow = Flow[ByteString]
-      .mapAsync(1)(data => {
-        val message = data.utf8String
-        Marshal(
-          ErrorResponse(
-            context = contextURL,
-            statusCode = statusCode,
-            description = message)).to[MessageEntity]
-      })
-      .flatMapConcat(_.dataBytes)
-
-    res
-      .transformEntityDataBytes(errorResponseMarshallingFlow)
-      .mapEntity(entity =>
-        entity.withContentType(ContentTypes.`application/json`))
-  }
 }
